refactor: InputView와 OutputView를 GameView로 합침
