+++ src/main/java/baseball/Application.java
@@ -3,11 +3,11 @@ package baseball;
 import static baseball.model.Status.REPLAY;
 
 import baseball.controller.GameController;
-import baseball.view.OutputView;
+import baseball.view.GameView;
 
 public class Application {
     public static void main(String[] args) {
-        OutputView.printStartMessage();
+        GameView.printStartMessage();
         GameController gameController = new GameController();
         do {
             gameController.run();
